Secured edition and deletion on backend

+++ backend/routes/posts.js
@@ -36,7 +36,8 @@ router.post('',
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId
     });
     post.save().then(createdPost => {
         res.status(201).json({
@@ -64,9 +65,13 @@ router.put('/:id',
         content: req.body.content,
         imagePath: imagePath
     })
-    Post.updateOne({_id: req.params.id}, post)
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post)
         .then(result => {
-            res.status(200).json({message: 'Update successful!'});
+            if (result.modifiedCount > 0) {
+                res.status(200).json({message: 'Update successful!'});
+            } else {
+                res.status(401).json({message: 'Not authorized!'});
+            }
         })
 })
 
@@ -107,9 +112,13 @@ router.get('/:id', (req,res,next) => {
 router.delete('/:id',
     checkAuth,
     (req,res,next) => {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-        res.status(200).json({message: 'Post deleted'});
-    })
-})
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+        if (result.deletedCount > 0) {
+            res.status(200).json({message: 'Deletion successful!'});
+        } else {
+            res.status(401).json({message: 'Not authorized!'});
+        }
+    });
+});
 
 module.exports = router;
